Remove all warnings raised by bugprone-* clang-tidy checks in acquisition adapters

+++ src/algorithms/acquisition/adapters/beidou_b1i_pcps_acquisition.cc
@@ -57,7 +57,7 @@ BeidouB1iPcpsAcquisition::BeidouB1iPcpsAcquisition(
             acq_parameters_.doppler_max = FLAGS_doppler_max;
         }
     doppler_max_ = acq_parameters_.doppler_max;
-    doppler_step_ = acq_parameters_.doppler_step;
+    doppler_step_ = static_cast<unsigned int>(acq_parameters_.doppler_step);
     fs_in_ = acq_parameters_.fs_in;
     item_type_ = acq_parameters_.item_type;
     item_size_ = acq_parameters_.it_size;
